fix: upload content with using latest API
